introduce gdml based example
