+++ Examples/Run/Geant4/DD4hep/CMakeLists.txt
@@ -1,8 +1,8 @@
 add_executable(
-  ActsExampleGdmlGeantinoRecording
-  GdmlGeantinoRecording.cpp)
+  ActsExampleDD4hepGeantinoRecording
+  DD4hepGeantinoRecording.cpp)
 target_link_libraries(
-  ActsExampleGdmlGeantinoRecording
+  ActsExampleDD4hepGeantinoRecording
   PRIVATE
     ActsExamplesCommon
     ActsExamplesDetectorDD4hep
@@ -13,7 +13,5 @@ target_link_libraries(
     Boost::program_options)
 
 install(
-  TARGETS ActsExampleGdmlGeantinoRecording
+  TARGETS ActsExampleDD4hepGeantinoRecording
   RUNTIME DESTINATION ${CMAKE_INSTALL_BINDIR})
-
-add_subdirectory_if(DD4hep ACTS_BUILD_EXAMPLES_DD4HEP)
